requirements.txt
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.166.0
python-dotenv==1.1.0
loguru==0.7.3
beautifulsoup4==4.13.3
openai==1.70.0
pytest==8.3.5
tiktoken==0.9.0
fastapi==0.115.12
uvicorn==0.34.0
itsdangerous==2.1.2
starlette==0.27.0